Prisma queries (#21)
* Add user photo generation function in seed.mjs

* Update getUsers function in prisma.ts

The getUsers function has been significantly expanded to provide more detailed information about jobseekers. It involves fields like the college enrollment status, highest level of study, job title, resume, work experience, and more. This update improves the depth of information that can be pulled from the PrismaClient for each jobseeker.

* Update JobSeekerCardView component in application UI

The JobSeekerCardView component has been updated to handle SkillDTO objects instead of simple strings for displaying user skills. The usage of SchoolDTO will provide more detailed and accurate depiction of the user's skills. Additionally, some sections such as GPA and seniority status are commented out to clean up display and better focus on the relevant information.

* Refactor Pill component to have a name

* Update property values in JobSeekerCardView component

The properties in JobSeekerCardView component have been updated to match the data structure of the incoming jobSeeker object. Refactored to map to the EmployerViewDTO object.

* Add new EmployerViewDTO

A new file, EmployerViewDTO.ts, has been added to the data/dtos directory. This file defines the structure of the EmployerViewDTO object which contains properties that match the upcoming changes to the jobSeeker data object.

* Add new SkillDTO type definition

Created a new file named SkillDTO.ts in the data/dtos directory. This file introduces the SkillDTO type, which includes unique id, skill name, and related info URL as its properties.

* Remove unused import in listview page

Removed the unused SkillDTO import in the Employer dashboard listview page. The import was not necessary because no properties from the SkillDTO type were used in the file.

* Fix: removed extra semicolon

* Rename EmployerViewDTO to JobSeekerCardViewDTO and trim fields

Renamed an exported type from EmployerViewDTO to JobSeekerCardViewDTO for relevance.

* Add new type definition 'JobseekerSkill'

Created a new type definition to match the result of the jobSeekerCardView prisma query

* Refactor SKillDTO

This commit flattens the SkillDTO by removing the 'skills' object, so it directly contains 'skill_id', 'skill_name', and 'skill_info_url'.

* Replace getUsers with getJobSeekerCardView in dashboard

* Refactor getUsers function to getJobSeekerCardView and add new fetching methods.

This commit mainly replaces the existing getUsers function with getJobSeekerCardView to fetch jobseekers data more specifically for the card view. It also provides ways to get jobseekers by their skills and a specific method to return a single jobseeker by id for employer view. Some fields from the original function are commented out to limit the information fetched for the card view display.

* Refactor display top three skills to top N Skills

I was thinking we may want to adjust this to show the top 5.

* Add top margin to Pill component

Added "mt-1" to the class of the button element in the Pill component. This will ensure a better visual appearance by providing some space between the top of the pill and any elements above it.

* Remove unnecessary import in JobSeekerCardViewDTO

* Correct commit:
Add console logging and return statement in prisma.ts

This commit includes the addition of a console logging statement for empView and a return statement within the same context. This was done for improved debugging and ensuring the correct data is being passed.

* Removed console.log

* Remove TODO comment from getJobSeekerEmployerView function

* WIP: DTO is not currently implemented but included for reference

This DTO needs types aligned with the database to before it can be used