Filtered search 2 (#24)
* implemented new query method

Implemented method for sanity check but need SearchBar component converted to a client rendered component and method used there.

* Refactor Prisma usage for job seeker data retrieval

Use the JobSeekerCardViewDTO, created a select variable to reduce code duplication. Modified prisma query to take in skills: [] and yearsWorkExp: number

* Set default values for job seeker query parameters

Assigned default values to the parameters 'skills' and 'yearsWorkExp' in the function 'getFilteredJobSeekerCardView'

* Set default values for job seeker query parameters

Assigned default values to the parameters 'skills' and 'yearsWorkExp' in the function 'getFilteredJobSeekerCardView'

* search bar now updates search term

* Update jobseeker page to fetch and display data from API

The Job Seeker page has been updated to fetch data from an API endpoint asynchronously. A new function fetchFilteredJobSeekerCardView has been added, and the component's state has been updated to handle fetched data and loading state.

* Add filtered route for work exp and skills in jobseekers API

A new route has been added to the jobseekers API to filter jobseeker data based on specific conditions such as work experience and skills. This route employs Prisma Client to fetch and filter data as per the specified parameters and returns the results as JSON. The route handler function is set to handle HTTP POST requests and appropriately address input request data.

* Export `jobSeekerCardViewSelect` from `prisma.ts`

This constant is used to select specific fields of jobseeker data and with this change, it can be used by other files/modules, providing a more modular approach to data selection.

* Refactor SearchBar functionality with `useState` and keyboard interaction

Added a `useState` to control the input value in the SearchBar. This change also includes functionality to perform a search on pressing the 'Enter' key.

* Remove unused import from Employer's Dashboard ListView

* Remove unused 'SetStateAction' import in SearchBar.tsx

---------

Co-authored-by: not-damien <93546644+not-damien@users.noreply.github.com>