Implements an API route for jobseeker profile introduction page
This can be used to both update and create new entries. Email must be provided to update an existing record.This new API handles the creation of a complete jobseeker profile by entering input page data into relevant tables in the database. Included is a transaction to ensure the atomicity of data input.