Update default constraints in several tables
In this commit, we are updating default constraints in several tables alongside adding 'edu_start_date' and 'edu_end_date' fields in the 'jobseekers' table. This procedure was performed by dropping the existing constraints and adding new defaults, with the methods applied to the 'case_mgmt', 'project_experiences', and 'skill_subcategories' tables.