Update education fields in jobseeker profile creation
The update adjusts the logic of date inputs and adds a new conditional field. The date selection fields now accept input in the form "YYYY-MM-01" and automatically insert the first day of the month for start and completion dates. It also introduces a new boolean field that marks whether the user is enrolled in college, depending on the education program selected.