fix: guarantee a null value on a property that maps to an Enum.
Originally the mapToEnum would return `undefined` if there wasn't a mapping. The Prisma ORM will omit this property if it is assigned undefined. The change to the mapToEnum function returns a null instead. Which will guarantee the property exists and be assigned a null value instead of being omitted from the json response object.