feat(eduProviders): update and create training provider programs (#feature/upsert-training-provider-programs)
Two optional fields, "description" and "targetedJobRoles", were added to ReadEduProviderProgramDetailDTO. The "targetedJobRoles" field in "ProviderProgramDetail" and "ProviderProgramDetailUpdate" was changed to an array of strings.

The update and create methods in the "upsertProviderProgramDetail" function were expanded. They now include all fields from the "programDetail" parameter. The "targetedJobRoles", "locations", "faq", and "pathways" fields are transformed to strings with a "~" delimiter or JSON strings where necessary.