feat(events-frontend): Replace 'inter' font with 'roboto' across the codebase (#feature/events-frontend)
In various files across the codebase, the 'inter' font has been replaced with the 'roboto' font. This change has been made for consistency and to improve the overall aesthetics of the application. Changes were made in the following files: layout.tsx (in the edit-profile/employer, layout, edit-profile/jobseeker directories), fonts.ts, and TagsWithAutocomplete.tsx. The weights for the 'roboto' font have also been defined in fonts.ts.