feat(careerPrep): enhance transactional handling of career preparation updates (#collect-other-priority-population-and-street-address)
This commit significantly enhances the transactional handling of career preparation updates. It includes a new function `updateCareerPrepStatusCardViewTx` which updates the career preparation status card view for a job seeker. It also corrects a typo in a comment regarding the `careerPrepAssessment` find operation. Furthermore, the `prisma.$transaction` callback now uses the `tx` transaction client for `updateCareerPrepEnrollment` and `updateCareerPrepStatusCardViewTx` to ensure transactional consistency. Lastly, the `prisma.careerPrepAssessment.update` function is now awaited to ensure completion.